pyproject.toml
[project]
name = "study-rag-chatbot-document-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "faiss-cpu>=1.13.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.7",
    "langchain-experimental>=0.4.0",
    "langchain-huggingface>=1.0.1",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "openai>=2.8.1",
    "pydantic>=2.12.4",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "wikipedia>=1.4.0",
]
